Dave PERRY:       (via Stuart BUCHANAN)
1. Make it more gentile and predictable in inverted flight, and
2. Allow solid predictable full stall landings (3 point landings).

With the rudder waggle technique, take-offs seem improved (no dragging
of the wing tips and nearly down the center of the runway).

Before this tweak, I could not do an inverted loop w/o having it do an
inverted snap roll out of the loop.  You still have to be smooth and
gentile with the stick when inverted.  I added aileron to the mstab to
get more positive response. The real sc1 did not have these but later
pitts did."